requirements.txt
accelerate
flask
numpy
pillow
pytest
torch
transformers

